requirements.txt
Django
psycopg2-binary
python-decouple
django-extensions
djangorestframework
markdown
django-filter
drf-yasg
Pillow
model_mommy
coverage